mv dirs.
